Refactor Pawn and Board, Add public enum Color
Pawn 클래스 내부에 있던 Representation enum을 제거하고 Color enum의 representation 변수를 만들어줘서 White('p')와 같은 형식으로
변수를 관리할 수 있게 변경했다.
클래스 간의 결합도를 줄이기 위해서 Pawn 클래스 내부에 있던 Color enum을 분리해서 public enum으로 변경했다.
Board 클래스의 getWhite(Black)PawnsResult() 메서드에서 forEach문을 사용하고 있었느데 이를 개선된 for loop문으로 변경했다.
