+++ src/main/java/net/sky/pieces/Color.java
@@ -1,0 +1,14 @@
+package net.sky.pieces;
+
+public enum Color {
+    WHITE('p'), BLACK('P');
+    private final char representation;
+
+    Color(char representation) {
+        this.representation = representation;
+    }
+
+    char getRepresentation() {
+        return this.representation;
+    }
+}
