+++ src/test/java/net/sky/pieces/PawnTest.java
@@ -3,8 +3,6 @@ package net.sky.pieces;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import net.sky.pieces.Pawn.Color;
-import net.sky.pieces.Pawn.Representation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,12 +12,12 @@ class PawnTest {
     @Test
     @DisplayName("폰 색깔 및 색깔별 출력방식 확인 테스트")
     void verifyColorTypeByPawn() {
-        verifyPawn(Color.WHITE, Representation.p);
-        verifyPawn(Color.BLACK, Representation.P);
+        verifyPawn(Color.WHITE, 'p');
+        verifyPawn(Color.BLACK, 'P');
     }
 
-    void verifyPawn(final Color color, final Representation representation) {
-        Pawn pawn = new Pawn(color, representation);
+    void verifyPawn(final Color color, final char representation) {
+        Pawn pawn = new Pawn(color);
         assertAll(
             () -> assertThat(pawn.getColor()).isEqualTo(color),
             () -> assertThat(pawn.getRepresentation()).isEqualTo(representation)
